% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erkpalettes.R
\docType{data}
\name{erkPalettes}
\alias{erkPalettes}
\title{Complete list of palettes.}
\format{
An object of class \code{list} of length 3.
}
\usage{
erkPalettes
}
\description{
Use names(erkPalettes) to return all possible palette names. Current choices are:
\code{erk}, \code{erksemi}, \code{erklight}.
Use \code{\link{erk.brewer}} to construct palettes.
}
\keyword{datasets}
