% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_erk}
\alias{theme_erk}
\title{ERK Theme}
\usage{
theme_erk(base_family = "sans", base_size = 10)
}
\arguments{
\item{base_family}{'sans', 'serif', 'mono' or any installed font, like 'Open Sans Light'}

\item{base_size}{numeric, defaults to 10 points.}
}
\value{
ggproto object
}
\description{
ERK Theme
}
